+++ ports/k210-freertos/mpy_support/omv/include/fb_alloc.h
@@ -11,13 +11,13 @@
 #include <stdint.h>
 void fb_alloc_fail();
 void fb_alloc_init0();
-uint32_t fb_avail();
+uint64_t fb_avail();
 void fb_alloc_mark();
 void fb_alloc_free_till_mark();
-void *fb_alloc(uint32_t size);
-void *fb_alloc0(uint32_t size);
-void *fb_alloc_all(uint32_t *size); // returns pointer and sets size
-void *fb_alloc0_all(uint32_t *size); // returns pointer and sets size
+void *fb_alloc(uint64_t size);
+void *fb_alloc0(uint64_t size);
+void *fb_alloc_all(uint64_t *size); // returns pointer and sets size
+void *fb_alloc0_all(uint64_t *size); // returns pointer and sets size
 void fb_free();
 void fb_free_all();
 #endif /* __FF_ALLOC_H__ */
